+++ prover/src/constraints/evaluator/mod.rs
@@ -13,6 +13,8 @@ pub use default::DefaultConstraintEvaluator;
 mod boundary;
 use boundary::BoundaryConstraints;
 
+mod lagrange;
+
 mod periodic_table;
 use periodic_table::PeriodicValueTable;
